created multiple layout components for different landing page and app, refactored navigation component to take in a 'route' prop
