--- app/layout.tsx
@@ -1,27 +1,0 @@
-import { GeistSans } from "geist/font/sans";
-import "@/app/globals.css";
-import ThemeProviderWrapper from "@/components/theme-provider-wrapper";
-
-const defaultUrl = process.env.VERCEL_URL
-  ? `https://${process.env.VERCEL_URL}`
-  : "http://localhost:3000";
-
-export const metadata = {
-  metadataBase: new URL(defaultUrl),
-  title: "Campus Connect",
-  description: "Your rental paradise",
-};
-
-export default function RootLayout({
-  children,
-}: {
-  children: React.ReactNode;
-}) {
-  return (
-    <html lang="en" className={GeistSans.className} suppressHydrationWarning>
-      <body className="min-h-screen flex flex-col bg-background text-foreground">
-        <ThemeProviderWrapper>{children}</ThemeProviderWrapper>
-      </body>
-    </html>
-  );
-}
